[Bronze V] Title: 꼬마 정민, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/11382. 꼬마 정민/꼬마 정민.cc
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
-int main()
+int main ()
 {
-    long long a, b, c;
-    
-    scanf("%lld %lld %lld", &a, &b, &c);
-    printf("%lld", a + b + c);
-    return 0;
+	long a,b,c;
+	
+	scanf("%ld %ld %ld", &a, &b, &c);
+	printf("%ld", a + b + c);
+	return 0;
 }
